Replace official Live Preview extension with Live Server
The official extension seems to have problems with the auto-generated files. It reloads the page randomly, even when there's no changes.

Ritwick Dey's extension seems to work better: it has an ignore files option and even a path mount option. I've still left the symlink code enabled just in case, but it should be working with the mount points defined in `.vscode/settings.json`.